+++ example/a.ts
@@ -1,13 +1,13 @@
 import { Elysia } from '../src'
 
-const a = (config = {}) =>
-	new Elysia({
-		name: 'a',
-		seed: config
-	}).get('/', () => 'a')
+const app = new Elysia({
+	cookie: {
+		httpOnly: true
+	}
+})
+	.get('/multiple', ({ cookie: { name, president } }) => {
+		name.value = 'Himari'
 
-const app = new Elysia().use(a()).listen(3000)
-
-await app.modules
-
-console.log(app.routes)
+		return 'ok'
+	})
+	.listen(3000)