+++ example/a/handler.ts
@@ -1,8 +1,0 @@
-import { Elysia } from '../../src'
-import { baseElysia, setup } from './main'
-
-export const usersHandler = new Elysia()
-    .use(setup)
-	.get('/users', ({ store: { count } }) => {
-		return `${count} users`
-	})